+++ sdk/bcs/tests/builder.test.ts
@@ -2,93 +2,89 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { describe, expect, test } from 'vitest';
-import { BcsType, BcsBuilder } from '../src/builder.js';
+import { bcs } from '../src/bcs.js';
 import { toHEX, BcsWriter, BcsReader, toB58 } from '../src';
+import { BcsType } from '../src/bcs-type.js';
 
-describe('BcsBuilder', () => {
+describe('bcs', () => {
 	describe('base types', () => {
-		testType('true', BcsBuilder.bool(), true, '01');
-		testType('false', BcsBuilder.bool(), false, '00');
-		testType('uleb128 0', BcsBuilder.uleb128(), 0, '00');
-		testType('uleb128 1', BcsBuilder.uleb128(), 1, '01');
-		testType('uleb128 127', BcsBuilder.uleb128(), 127, '7f');
-		testType('uleb128 128', BcsBuilder.uleb128(), 128, '8001');
-		testType('uleb128 255', BcsBuilder.uleb128(), 255, 'ff01');
-		testType('uleb128 256', BcsBuilder.uleb128(), 256, '8002');
-		testType('uleb128 16383', BcsBuilder.uleb128(), 16383, 'ff7f');
-		testType('uleb128 16384', BcsBuilder.uleb128(), 16384, '808001');
-		testType('uleb128 2097151', BcsBuilder.uleb128(), 2097151, 'ffff7f');
-		testType('uleb128 2097152', BcsBuilder.uleb128(), 2097152, '80808001');
-		testType('uleb128 268435455', BcsBuilder.uleb128(), 268435455, 'ffffff7f');
-		testType('uleb128 268435456', BcsBuilder.uleb128(), 268435456, '8080808001');
-		testType('u8 0', BcsBuilder.u8(), 0, '00');
-		testType('u8 1', BcsBuilder.u8(), 1, '01');
-		testType('u8 255', BcsBuilder.u8(), 255, 'ff');
-		testType('u16 0', BcsBuilder.u16(), 0, '0000');
-		testType('u16 1', BcsBuilder.u16(), 1, '0100');
-		testType('u16 255', BcsBuilder.u16(), 255, 'ff00');
-		testType('u16 256', BcsBuilder.u16(), 256, '0001');
-		testType('u16 65535', BcsBuilder.u16(), 65535, 'ffff');
-		testType('u32 0', BcsBuilder.u32(), 0, '00000000');
-		testType('u32 1', BcsBuilder.u32(), 1, '01000000');
-		testType('u32 255', BcsBuilder.u32(), 255, 'ff000000');
-		testType('u32 256', BcsBuilder.u32(), 256, '00010000');
-		testType('u32 65535', BcsBuilder.u32(), 65535, 'ffff0000');
-		testType('u32 65536', BcsBuilder.u32(), 65536, '00000100');
-		testType('u32 16777215', BcsBuilder.u32(), 16777215, 'ffffff00');
-		testType('u32 16777216', BcsBuilder.u32(), 16777216, '00000001');
-		testType('u32 4294967295', BcsBuilder.u32(), 4294967295, 'ffffffff');
-		testType('u64 0', BcsBuilder.u64(), 0, '0000000000000000', 0n);
-		testType('u64 1', BcsBuilder.u64(), 1, '0100000000000000', 1n);
-		testType('u64 255', BcsBuilder.u64(), 255n, 'ff00000000000000');
-		testType('u64 256', BcsBuilder.u64(), 256n, '0001000000000000');
-		testType('u64 65535', BcsBuilder.u64(), 65535n, 'ffff000000000000');
-		testType('u64 65536', BcsBuilder.u64(), 65536n, '0000010000000000');
-		testType('u64 16777215', BcsBuilder.u64(), 16777215n, 'ffffff0000000000');
-		testType('u64 16777216', BcsBuilder.u64(), 16777216n, '0000000100000000');
-		testType('u64 4294967295', BcsBuilder.u64(), 4294967295n, 'ffffffff00000000');
-		testType('u64 4294967296', BcsBuilder.u64(), 4294967296n, '0000000001000000');
-		testType('u64 1099511627775', BcsBuilder.u64(), 1099511627775n, 'ffffffffff000000');
-		testType('u64 1099511627776', BcsBuilder.u64(), 1099511627776n, '0000000000010000');
-		testType('u64 281474976710655', BcsBuilder.u64(), 281474976710655n, 'ffffffffffff0000');
-		testType('u64 281474976710656', BcsBuilder.u64(), 281474976710656n, '0000000000000100');
-		testType('u64 72057594037927935', BcsBuilder.u64(), 72057594037927935n, 'ffffffffffffff00');
-		testType('u64 72057594037927936', BcsBuilder.u64(), 72057594037927936n, '0000000000000001');
-		testType(
-			'u64 18446744073709551615',
-			BcsBuilder.u64(),
-			18446744073709551615n,
-			'ffffffffffffffff',
-		);
-		testType('u128 0', BcsBuilder.u128(), 0n, '00000000000000000000000000000000');
-		testType('u128 1', BcsBuilder.u128(), 1n, '01000000000000000000000000000000');
-		testType('u128 255', BcsBuilder.u128(), 255n, 'ff000000000000000000000000000000');
+		testType('true', bcs.bool(), true, '01');
+		testType('false', bcs.bool(), false, '00');
+		testType('uleb128 0', bcs.uleb128(), 0, '00');
+		testType('uleb128 1', bcs.uleb128(), 1, '01');
+		testType('uleb128 127', bcs.uleb128(), 127, '7f');
+		testType('uleb128 128', bcs.uleb128(), 128, '8001');
+		testType('uleb128 255', bcs.uleb128(), 255, 'ff01');
+		testType('uleb128 256', bcs.uleb128(), 256, '8002');
+		testType('uleb128 16383', bcs.uleb128(), 16383, 'ff7f');
+		testType('uleb128 16384', bcs.uleb128(), 16384, '808001');
+		testType('uleb128 2097151', bcs.uleb128(), 2097151, 'ffff7f');
+		testType('uleb128 2097152', bcs.uleb128(), 2097152, '80808001');
+		testType('uleb128 268435455', bcs.uleb128(), 268435455, 'ffffff7f');
+		testType('uleb128 268435456', bcs.uleb128(), 268435456, '8080808001');
+		testType('u8 0', bcs.u8(), 0, '00');
+		testType('u8 1', bcs.u8(), 1, '01');
+		testType('u8 255', bcs.u8(), 255, 'ff');
+		testType('u16 0', bcs.u16(), 0, '0000');
+		testType('u16 1', bcs.u16(), 1, '0100');
+		testType('u16 255', bcs.u16(), 255, 'ff00');
+		testType('u16 256', bcs.u16(), 256, '0001');
+		testType('u16 65535', bcs.u16(), 65535, 'ffff');
+		testType('u32 0', bcs.u32(), 0, '00000000');
+		testType('u32 1', bcs.u32(), 1, '01000000');
+		testType('u32 255', bcs.u32(), 255, 'ff000000');
+		testType('u32 256', bcs.u32(), 256, '00010000');
+		testType('u32 65535', bcs.u32(), 65535, 'ffff0000');
+		testType('u32 65536', bcs.u32(), 65536, '00000100');
+		testType('u32 16777215', bcs.u32(), 16777215, 'ffffff00');
+		testType('u32 16777216', bcs.u32(), 16777216, '00000001');
+		testType('u32 4294967295', bcs.u32(), 4294967295, 'ffffffff');
+		testType('u64 0', bcs.u64(), 0, '0000000000000000', 0n);
+		testType('u64 1', bcs.u64(), 1, '0100000000000000', 1n);
+		testType('u64 255', bcs.u64(), 255n, 'ff00000000000000');
+		testType('u64 256', bcs.u64(), 256n, '0001000000000000');
+		testType('u64 65535', bcs.u64(), 65535n, 'ffff000000000000');
+		testType('u64 65536', bcs.u64(), 65536n, '0000010000000000');
+		testType('u64 16777215', bcs.u64(), 16777215n, 'ffffff0000000000');
+		testType('u64 16777216', bcs.u64(), 16777216n, '0000000100000000');
+		testType('u64 4294967295', bcs.u64(), 4294967295n, 'ffffffff00000000');
+		testType('u64 4294967296', bcs.u64(), 4294967296n, '0000000001000000');
+		testType('u64 1099511627775', bcs.u64(), 1099511627775n, 'ffffffffff000000');
+		testType('u64 1099511627776', bcs.u64(), 1099511627776n, '0000000000010000');
+		testType('u64 281474976710655', bcs.u64(), 281474976710655n, 'ffffffffffff0000');
+		testType('u64 281474976710656', bcs.u64(), 281474976710656n, '0000000000000100');
+		testType('u64 72057594037927935', bcs.u64(), 72057594037927935n, 'ffffffffffffff00');
+		testType('u64 72057594037927936', bcs.u64(), 72057594037927936n, '0000000000000001');
+		testType('u64 18446744073709551615', bcs.u64(), 18446744073709551615n, 'ffffffffffffffff');
+		testType('u128 0', bcs.u128(), 0n, '00000000000000000000000000000000');
+		testType('u128 1', bcs.u128(), 1n, '01000000000000000000000000000000');
+		testType('u128 255', bcs.u128(), 255n, 'ff000000000000000000000000000000');
 		testType(
 			'u128 18446744073709551615',
-			BcsBuilder.u128(),
+			bcs.u128(),
 			18446744073709551615n,
 			'ffffffffffffffff0000000000000000',
 		);
 		testType(
 			'u128 18446744073709551615',
-			BcsBuilder.u128(),
+			bcs.u128(),
 			18446744073709551616n,
 			'00000000000000000100000000000000',
 		);
 		testType(
 			'u128 340282366920938463463374607431768211455',
-			BcsBuilder.u128(),
+			bcs.u128(),
 			340282366920938463463374607431768211455n,
 			'ffffffffffffffffffffffffffffffff',
 		);
 	});
 
 	describe('vector', () => {
-		testType('vector([])', BcsBuilder.vector(BcsBuilder.u8()), [], '00');
-		testType('vector([1, 2, 3])', BcsBuilder.vector(BcsBuilder.u8()), [1, 2, 3], '03010203');
+		testType('vector([])', bcs.vector(bcs.u8()), [], '00');
+		testType('vector([1, 2, 3])', bcs.vector(bcs.u8()), [1, 2, 3], '03010203');
 		testType(
 			'vector([1, null, 3])',
-			BcsBuilder.vector(BcsBuilder.option(BcsBuilder.u8())),
+			bcs.vector(bcs.option(bcs.u8())),
 			[1, null, 3],
 			// eslint-disable-next-line no-useless-concat
 			'03' + '0101' + '00' + '0103',
@@ -96,11 +92,11 @@ describe('BcsBuilder', () => {
 	});
 
 	describe('fixedVector', () => {
-		testType('fixedVector([])', BcsBuilder.fixedVector(0, BcsBuilder.u8()), [], '');
-		testType('vector([1, 2, 3])', BcsBuilder.fixedVector(3, BcsBuilder.u8()), [1, 2, 3], '010203');
+		testType('fixedVector([])', bcs.array(0, bcs.u8()), [], '');
+		testType('vector([1, 2, 3])', bcs.array(3, bcs.u8()), [1, 2, 3], '010203');
 		testType(
 			'fixedVector([1, null, 3])',
-			BcsBuilder.fixedVector(3, BcsBuilder.option(BcsBuilder.u8())),
+			bcs.array(3, bcs.option(bcs.u8())),
 			[1, null, 3],
 			// eslint-disable-next-line no-useless-concat
 			'0101' + '00' + '0103',
@@ -108,18 +104,13 @@ describe('BcsBuilder', () => {
 	});
 
 	describe('options', () => {
-		testType('optional u8 undefined', BcsBuilder.option(BcsBuilder.u8()), undefined, '00', null);
-		testType('optional u8 null', BcsBuilder.option(BcsBuilder.u8()), null, '00');
-		testType('optional u8 0', BcsBuilder.option(BcsBuilder.u8()), 0, '0100');
-		testType(
-			'optional vector(null)',
-			BcsBuilder.option(BcsBuilder.vector(BcsBuilder.u8())),
-			null,
-			'00',
-		);
+		testType('optional u8 undefined', bcs.option(bcs.u8()), undefined, '00', null);
+		testType('optional u8 null', bcs.option(bcs.u8()), null, '00');
+		testType('optional u8 0', bcs.option(bcs.u8()), 0, '0100');
+		testType('optional vector(null)', bcs.option(bcs.vector(bcs.u8())), null, '00');
 		testType(
 			'optional vector([1, 2, 3])',
-			BcsBuilder.option(BcsBuilder.vector(BcsBuilder.option(BcsBuilder.u8()))),
+			bcs.option(bcs.vector(bcs.option(bcs.u8()))),
 			[1, null, 3],
 			// eslint-disable-next-line no-useless-concat
 			'01' + '03' + '0101' + '00' + '0103',
@@ -127,60 +118,61 @@ describe('BcsBuilder', () => {
 	});
 
 	describe('string', () => {
-		testType('string empty', BcsBuilder.string(), '', '00');
-		testType('string hello', BcsBuilder.string(), 'hello', '0568656c6c6f');
+		testType('string empty', bcs.string(), '', '00');
+		testType('string hello', bcs.string(), 'hello', '0568656c6c6f');
 		testType(
 			'string çå∞≠¢õß∂ƒ∫',
-			BcsBuilder.string(),
+			bcs.string(),
 			'çå∞≠¢õß∂ƒ∫',
 			'18c3a7c3a5e2889ee289a0c2a2c3b5c39fe28882c692e288ab',
 		);
 	});
 
 	describe('bytes', () => {
-		testType('bytes', BcsBuilder.bytes(4), new Uint8Array([1, 2, 3, 4]), '01020304');
+		testType('bytes', bcs.bytes(4), new Uint8Array([1, 2, 3, 4]), '01020304');
 	});
 
 	describe('hex', () => {
-		testType('hex', BcsBuilder.hex(), '01020304', '0401020304');
+		testType('hex', bcs.hex(), '01020304', '0401020304');
 	});
 
 	describe('base64', () => {
-		testType('base64', BcsBuilder.base64(), 'AQIDBA==', '0401020304');
+		testType('base64', bcs.base64(), 'AQIDBA==', '0401020304');
 	});
 
 	describe('b58', () => {
-		testType('b58', BcsBuilder.base58(), toB58(new Uint8Array([1, 2, 3, 4])), '0401020304');
+		testType('b58', bcs.base58(), toB58(new Uint8Array([1, 2, 3, 4])), '0401020304');
 	});
 
 	describe('tuples', () => {
-		testType('tuple(u8, u8)', BcsBuilder.tuple([BcsBuilder.u8(), BcsBuilder.u8()]), [1, 2], '0102');
+		testType('tuple(u8, u8)', bcs.tuple([bcs.u8(), bcs.u8()]), [1, 2], '0102');
 		testType(
 			'tuple(u8, string, boolean)',
-			BcsBuilder.tuple([BcsBuilder.u8(), BcsBuilder.string(), BcsBuilder.bool()]),
+			bcs.tuple([bcs.u8(), bcs.string(), bcs.bool()]),
 			[1, 'hello', true],
 			'010568656c6c6f01',
 		);
 
 		testType(
 			'tuple(null, u8)',
-			BcsBuilder.tuple([BcsBuilder.option(BcsBuilder.u8()), BcsBuilder.option(BcsBuilder.u8())]),
+			bcs.tuple([bcs.option(bcs.u8()), bcs.option(bcs.u8())]),
 			[null, 1],
 			'000101',
 		);
 	});
 
 	describe('structs', () => {
-		const MyStruct = BcsBuilder.struct({
-			boolean: BcsBuilder.bool(),
-			bytes: BcsBuilder.vector(BcsBuilder.u8()),
-			label: BcsBuilder.string(),
+		const MyStruct = bcs.struct('MyStruct', {
+			boolean: bcs.bool(),
+			bytes: bcs.vector(bcs.u8()),
+			label: bcs.string(),
 		});
 
-		const Wrapper = BcsBuilder.struct({
+		const Wrapper = bcs.struct('Wrapper', {
 			inner: MyStruct,
-			name: BcsBuilder.string(),
+			name: bcs.string(),
 		});
+
 		testType(
 			'struct { boolean: bool, bytes: Vec<u8>, label: String }',
 			MyStruct,
@@ -222,10 +214,10 @@ describe('BcsBuilder', () => {
 	});
 
 	describe('enums', () => {
-		const E = BcsBuilder.enum({
-			Variant0: BcsBuilder.u16(),
-			Variant1: BcsBuilder.u8(),
-			Variant2: BcsBuilder.string(),
+		const E = bcs.enum('E', {
+			Variant0: bcs.u16(),
+			Variant1: bcs.u8(),
+			Variant2: bcs.string(),
 		});
 
 		testType('Enum::Variant0(1)', E, { Variant0: 1 }, '000100');
