+++ sdk/typescript/src/builder/TransactionBlockData.ts
@@ -18,11 +18,11 @@ import {
 } from 'superstruct';
 import { hashTypedData } from './hash.js';
 import { SuiObjectRef } from '../types/index.js';
-import { builder } from './bcs.js';
 import { TransactionType, TransactionBlockInput } from './Transactions.js';
 import { BuilderCallArg, PureCallArg } from './Inputs.js';
 import { create } from './utils.js';
 import { normalizeSuiAddress } from '../utils/sui-types.js';
+import { BCS } from '../bcs/index.js';
 
 export const TransactionExpiration = optional(
 	nullable(
@@ -31,7 +31,7 @@ export const TransactionExpiration = optional(
 );
 export type TransactionExpiration = Infer<typeof TransactionExpiration>;
 
-const StringEncodedBigint = define<string>('StringEncodedBigint', (val) => {
+const StringEncodedBigint = define<string | number | bigint>('StringEncodedBigint', (val) => {
 	if (!['string', 'number', 'bigint'].includes(typeof val)) return false;
 
 	try {
@@ -66,8 +66,8 @@ function prepareSuiAddress(address: string) {
 
 export class TransactionBlockDataBuilder {
 	static fromKindBytes(bytes: Uint8Array) {
-		const kind = builder.de('TransactionKind', bytes);
-		const programmableTx = kind?.ProgrammableTransaction;
+		const kind = BCS.TransactionKind.parse(bytes);
+		const programmableTx = 'ProgrammableTransaction' in kind ? kind.ProgrammableTransaction : null;
 		if (!programmableTx) {
 			throw new Error('Unable to deserialize from bytes.');
 		}
@@ -96,9 +96,10 @@ export class TransactionBlockDataBuilder {
 	}
 
 	static fromBytes(bytes: Uint8Array) {
-		const rawData = builder.de('TransactionData', bytes);
+		const rawData = BCS.TransactionData.parse(bytes);
 		const data = rawData?.V1;
-		const programmableTx = data?.kind?.ProgrammableTransaction;
+		const programmableTx =
+			'ProgrammableTransaction' in data.kind ? data?.kind?.ProgrammableTransaction : null;
 		if (!data || !programmableTx) {
 			throw new Error('Unable to deserialize from bytes.');
 		}
@@ -184,7 +185,7 @@ export class TransactionBlockDataBuilder {
 		};
 
 		if (onlyTransactionKind) {
-			return builder.ser('TransactionKind', kind, { maxSize: maxSizeBytes }).toBytes();
+			return BCS.TransactionKind.serialize(kind, { maxSize: maxSizeBytes });
 		}
 
 		const expiration = overrides?.expiration ?? this.expiration;
@@ -224,9 +225,7 @@ export class TransactionBlockDataBuilder {
 			},
 		};
 
-		return builder
-			.ser('TransactionData', { V1: transactionData }, { maxSize: maxSizeBytes })
-			.toBytes();
+		return BCS.TransactionData.serialize({ V1: transactionData }, { maxSize: maxSizeBytes });
 	}
 
 	getDigest() {
