+++ sdk/typescript/src/builder/Transactions.ts
@@ -22,6 +22,7 @@ import type { WellKnownEncoding } from './utils.js';
 import { TRANSACTION_TYPE, create } from './utils.js';
 import { TypeTagSerializer } from '../bcs/type-tag-serializer.js';
 import { normalizeSuiObjectId } from '../utils/sui-types.js';
+import type { TypeTag } from '../bcs/index.js';
 
 const option = <T extends Struct<any, any>>(some: T) =>
 	union([object({ None: union([literal(true), literal(null)]) }), object({ Some: some })]);
@@ -98,7 +99,9 @@ export const MakeMoveVecTransaction = object({
 	// TODO: ideally we should use `TypeTag` instead of `record()` here,
 	// but TypeTag is recursively defined and it's tricky to define a
 	// recursive struct in superstruct
-	type: optional(option(record(string(), unknown()))),
+	type: optional(option(record(string(), unknown()))) as never as Struct<
+		{ Some: TypeTag } | { None: true | null }
+	>,
 	objects: array(ObjectTransactionArgument),
 });
 export type MakeMoveVecTransaction = Infer<typeof MakeMoveVecTransaction>;
