+++ sdk/typescript/src/builder/index.ts
@@ -4,5 +4,5 @@
 export * from './TransactionBlock.js';
 export * from './Transactions.js';
 export * from './Inputs.js';
-export * from './bcs.js';
 export * from './serializer.js';
+export * from './bcs.js';
