+++ sdk/typescript/src/cryptography/publickey.ts
@@ -85,7 +85,7 @@ export abstract class PublicKey {
 		signature: Uint8Array | SerializedSignature,
 	): Promise<boolean> {
 		return this.verifyWithIntent(
-			bcs.ser(['vector', 'u8'], message).toBytes(),
+			bcs.vector(bcs.u8()).serialize(message),
 			signature,
 			IntentScope.PersonalMessage,
 		);
