+++ sdk/typescript/src/cryptography/signature.ts
@@ -4,7 +4,7 @@
 import { fromB64, toB64 } from '@mysten/bcs';
 import type { PublicKey } from './publickey.js';
 import type { MultiSigStruct } from '../multisig/publickey.js';
-import { builder } from '../builder/bcs.js';
+import { BCS } from '../bcs/index.js';
 
 export type SignatureScheme = 'ED25519' | 'Secp256k1' | 'Secp256r1' | 'MultiSig';
 
@@ -82,7 +82,7 @@ export function parseSerializedSignature(serializedSignature: SerializedSignatur
 		SIGNATURE_FLAG_TO_SCHEME[bytes[0] as keyof typeof SIGNATURE_FLAG_TO_SCHEME];
 
 	if (signatureScheme === 'MultiSig') {
-		const multisig: MultiSigStruct = builder.de('MultiSig', bytes.slice(1));
+		const multisig: MultiSigStruct = BCS.MultiSig.parse(bytes.slice(1));
 		return {
 			serializedSignature,
 			signatureScheme,
