+++ sdk/typescript/src/multisig/publickey.ts
@@ -16,9 +16,9 @@ import {
 	parseSerializedSignature,
 } from '../cryptography/signature.js';
 import { normalizeSuiAddress } from '../utils/sui-types.js';
-import { builder } from '../builder/bcs.js';
 // eslint-disable-next-line import/no-cycle
 import { publicKeyFromRawBytes } from '../verify/index.js';
+import { BCS, bcs } from '../bcs/index.js';
 
 type CompressedSignature =
 	| { ED25519: number[] }
@@ -75,13 +75,14 @@ export class MultiSigPublicKey extends PublicKey {
 
 		if (typeof value === 'string') {
 			this.rawBytes = fromB64(value);
-			this.multisigPublicKey = builder.de('MultiSigPublicKey', this.rawBytes);
+
+			this.multisigPublicKey = BCS.MultiSigPublicKey.parse(this.rawBytes);
 		} else if (value instanceof Uint8Array) {
 			this.rawBytes = value;
-			this.multisigPublicKey = builder.de('MultiSigPublicKey', this.rawBytes);
+			this.multisigPublicKey = BCS.MultiSigPublicKey.parse(this.rawBytes);
 		} else {
 			this.multisigPublicKey = value;
-			this.rawBytes = builder.ser('MultiSigPublicKey', value).toBytes();
+			this.rawBytes = BCS.MultiSigPublicKey.serialize(value);
 		}
 		if (this.multisigPublicKey.threshold < 1) {
 			throw new Error('Invalid threshold');
@@ -173,7 +174,7 @@ export class MultiSigPublicKey extends PublicKey {
 		const tmp = new Uint8Array(maxLength);
 		tmp.set([SIGNATURE_SCHEME_TO_FLAG['MultiSig']]);
 
-		tmp.set(builder.ser('u16', this.multisigPublicKey.threshold).toBytes(), 1);
+		tmp.set(bcs.u16().serialize(this.multisigPublicKey.threshold), 1);
 		// The initial value 3 ensures that following data will be after the flag byte and threshold bytes
 		let i = 3;
 		for (const { publicKey, weight } of this.publicKeys) {
@@ -207,8 +208,8 @@ export class MultiSigPublicKey extends PublicKey {
 
 		if (
 			!bytesEqual(
-				builder.ser('MultiSigPublicKey', this.multisigPublicKey).toBytes(),
-				builder.ser('MultiSigPublicKey', multisig.multisig_pk).toBytes(),
+				BCS.MultiSigPublicKey.serialize(this.multisigPublicKey),
+				BCS.MultiSigPublicKey.serialize(multisig.multisig_pk),
 			)
 		) {
 			return false;
@@ -279,7 +280,7 @@ export class MultiSigPublicKey extends PublicKey {
 			multisig_pk: this.multisigPublicKey,
 		};
 
-		const bytes = builder.ser('MultiSig', multisig).toBytes();
+		const bytes = BCS.MultiSig.serialize(multisig);
 		let tmp = new Uint8Array(bytes.length + 1);
 		tmp.set([SIGNATURE_SCHEME_TO_FLAG['MultiSig']]);
 		tmp.set(bytes, 1);
