+++ sdk/typescript/src/types/objects.ts
@@ -18,6 +18,7 @@ import {
 	nullable,
 	tuple,
 	unknown,
+	bigint,
 } from 'superstruct';
 import { ObjectOwner } from './common.js';
 import type { OwnedObjectRef } from './transactions.js';
@@ -31,7 +32,7 @@ export const SuiObjectRef = object({
 	/** Hex code as string representing the object id */
 	objectId: string(),
 	/** Object version */
-	version: union([number(), string()]),
+	version: union([number(), string(), bigint()]),
 });
 export type SuiObjectRef = Infer<typeof SuiObjectRef>;
 
@@ -283,7 +284,7 @@ export function getObjectId(data: SuiObjectResponse | SuiObjectRef | OwnedObject
 
 export function getObjectVersion(
 	data: SuiObjectResponse | SuiObjectRef | SuiObjectData,
-): string | number | undefined {
+): string | number | bigint | undefined {
 	if ('version' in data) {
 		return data.version;
 	}
