+++ sdk/typescript/src/utils/verify.ts
@@ -24,7 +24,7 @@ export async function verifyMessage(
 	if (scope === IntentScope.PersonalMessage) {
 		const messageBytes = messageWithIntent(
 			scope,
-			bcs.ser(['vector', 'u8'], typeof message === 'string' ? fromB64(message) : message).toBytes(),
+			bcs.vector(bcs.u8()).serialize(typeof message === 'string' ? fromB64(message) : message),
 		);
 
 		if (await signature.pubKey.verify(blake2b(messageBytes, { dkLen: 32 }), signature.signature)) {
